@teshull 2020-07-28
I merged this yesterday, but didn't see your new commit before merging. Also, I changed the other initializations within `AArch64GraphBuilderPlugins.registerUnsafePlugins` to use UnsafeAcessPlugin.

Looking at your new commit, I agree that the unaligned plugins should be registered in a different method. Can you rebase that commit on the current master, and also rename it to `registerUnsafeUnalignedPlugins` or something like that (I don't care for the registerInternalUnsafeSpecificPlugins name)?